refactor: extract the query surface into a shared component
The apex demo page had the query surface wired inline: the form, the
side-by-side comparison results, the executed-notebook results, and the
two streaming hooks that drive them. Move that whole surface into
`src/components/shared/QuerySurface.tsx` and mount it from
`(marketing)/page.tsx` in exactly today's configuration, so the surface
can later be mounted somewhere else without duplicating the wiring.

Behaviorally inert: every state variable, handler, and JSX block moves
verbatim. The page's framing content — hero copy, the "Go deeper" CTA,
the positioning band — stays with the page; the hero renders through the
component's `children` slot so it keeps its position inside the form's
container and the rendered DOM is unchanged.

Verified: a normalized DOM capture of `/` before and after is
byte-for-byte identical (the only raw differences are bundle chunk names
and the RSC flight payload). Lint, build, and tests all pass.
